django-uas/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
paho-mqtt==1.6.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
